newdjango/codetest/requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.12.2
mysqlclient==2.0.1
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
